added test for fft and edited previous tests
add a test of the new fft methods and compares the to the previous way of doing it and doing the same thing in different ways with the new methods, to make sure they all give the same answer.

Also edited test_file_selection_methods so that it loads from local, not a google bucket. Just to speed things up a bit.

All test run in 30s on my local machine.