fully working ezz variance calculations
corrects a few errors. Two errors described here https://github.com/ldeo-glaciology/xapres/issues/.
An additional error was that we were taking the sqrt of the variance and still calling it variance. This removes the sqrt within `weighted_least_squares`.
It also allows the fit to be performed on dask arrays.